Updated game for stun
Stun variable being saved to .mob files was causing the game to fail to load.

db.c had a clause which limited the number of values on a line in the file to 9 total.  this has been changed to 10 to accommodate for the new value added, GET_STUN.

All old mobs from the previous zones have been removed to ensure compatibility.